Logging: replace required Commons Logging with optional SLF4J, falling back to Java Util logging
This patch was inspired by @spmiller's work on a pluggable logging interface.  This patch caters to
two main audiences: those two don't want any extraneous dependencies (by removing the need for
commons logging) and those who want to be able to redirect and/or suppress logging via standard
logging configuration (when SLF4J and a binding are provided).

https://github.com/marceloverdijk/lesscss-java/pull/16